refactoring: basic optimization
