+++ src/components/Search/index.tsx
@@ -25,4 +25,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default Search;
+export default React.memo(Search);
